move Message object as a private class
